Move remove() to API module
